Fix Off By One Data Errors
This problem was caused by the assumption that requested data could be
read on the next clock. It actually takes two cycles in order to read
values from the memory unit due to its registered output. It was only
working before because of the read operation which is done after each
cycle to check for the halted state.

This change adds additional wait states to the ICP unit and the top
level memory/register bus to prevent off by one memory errors. The
program output is now correctly located at position 0.